Move Computation to own file
